Made AddressBookIterator an inner class. Removed AddressBook.toConfigText()
