+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/state/address/AddressBookInitializer.java
@@ -5,6 +5,7 @@ import static com.swirlds.logging.legacy.LogMarker.EXCEPTION;
 import static com.swirlds.logging.legacy.LogMarker.STARTUP;
 import static com.swirlds.platform.roster.RosterRetriever.retrieveActiveOrGenesisRoster;
 import static com.swirlds.platform.roster.RosterUtils.buildAddressBook;
+import static com.swirlds.platform.system.address.AddressBookUtils.addressBookConfigText;
 
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.platform.config.AddressBookConfig;
@@ -298,10 +299,10 @@ public class AddressBookInitializer {
                     .toFile();
             try (final FileWriter out = new FileWriter(debugFile)) {
                 out.write(CONFIG_ADDRESS_BOOK_HEADER + "\n");
-                out.write(configAddressBook.toConfigText() + "\n\n");
+                out.write(addressBookConfigText(configAddressBook) + "\n\n");
                 out.write(STATE_ADDRESS_BOOK_HEADER + "\n");
                 final String text =
-                        stateAddressBook == null ? STATE_ADDRESS_BOOK_NULL : stateAddressBook.toConfigText();
+                        stateAddressBook == null ? STATE_ADDRESS_BOOK_NULL : addressBookConfigText(stateAddressBook);
                 out.write(text + "\n\n");
                 out.write(USED_ADDRESS_BOOK_HEADER + "\n");
                 if (usedAddressBook == configAddressBook) {
@@ -309,14 +310,14 @@ public class AddressBookInitializer {
                 } else if (usedAddressBook == stateAddressBook) {
                     out.write(STATE_ADDRESS_BOOK_USED);
                 } else {
-                    out.write(usedAddressBook.toConfigText());
+                    out.write(addressBookConfigText(usedAddressBook));
                 }
                 out.write("\n\n");
             }
             final File usedFile = Path.of(this.pathToAddressBookDirectory.toString(), addressBookFileName)
                     .toFile();
             try (final FileWriter out = new FileWriter(usedFile)) {
-                out.write(usedAddressBook.toConfigText());
+                out.write(addressBookConfigText(usedAddressBook));
             }
         } catch (final IOException e) {
             logger.error(EXCEPTION.getMarker(), "Not able to write address book to file. ", e);
